+++ doc/changelog.rst
@@ -12,6 +12,7 @@ Unrealeased
 Core
 ~~~~
 
+- Fix bug in `Order.can_set_complete`
 - Currencies can be now created and edited through admin.
 
 Localization